backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
channels==4.0.0
asgiref==3.5.2
uvicorn==0.29.0